# public/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.35
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/v1/models/catalog:
    get:
      tags: [Models]
      summary: Harness capability catalog (Layer 3)
      description: >-
        Fork(harness B1). Agent-facing retrieval over the model tag index:
        every catalog model with categories, composite benchmark, per-axis
        benchmark scores (swe_bench, humaneval, math500, gpqa, mmlu,
        lmarena_elo), capability flags and context length. API-key policy
        (not management auth). Filters: category, axis, providers,
        minBenchmark, requireTools, requireVision, limit, offset.
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          required: false
        - name: axis
          in: query
          schema:
            type: string
            enum: [mmlu, humaneval, swe_bench, math500, gpqa, lmarena_elo]
          required: false
        - name: providers
          in: query
          schema:
            type: string
          required: false
        - name: minBenchmark
          in: query
          schema:
            type: number
          required: false
        - name: requireTools
          in: query
          schema:
            type: boolean
          required: false
        - name: requireVision
          in: query
          schema:
            type: boolean
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          required: false
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          required: false
      responses:
        "200":
          description: Capability catalog page
        "400":
          description: Invalid filter value
        "401":
          description: Unauthorized
  /api/v1/models/best:
    get:
      tags: [Models]
      summary: Best models for a task type (axis-ranked)
      description: >-
        Fork(harness B1). The allocator query: task uses the classifier
        vocabulary (code, research, math, vision, search, chat, image_gen)
        and maps through TASK_TYPE_TO_QUERY — the same mapping the classifier
        output and capability aliases use. Ranking is axis-aware: code sorts
        by SWE-bench, math by MATH-500, reasoning by GPQA, chat by LMArena
        ELO. Raw form: category= plus optional axis=. Distinct models,
        provider-diverse, with the task type's category fallback.
      security:
        - BearerAuth: []
      parameters:
        - name: task
          in: query
          schema:
            type: string
            enum: [code, research, math, reasoning, vision, search, chat, image_gen]
          required: false
        - name: category
          in: query
          schema:
            type: string
          required: false
        - name: axis
          in: query
          schema:
            type: string
            enum: [mmlu, humaneval, swe_bench, math500, gpqa, lmarena_elo]
          required: false
        - name: minBenchmark
          in: query
          schema:
            type: number
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
          required: false
      responses:
        "200":
          description: Ranked specialists for the task type
        "400":
          description: Missing/invalid task or category
        "401":
          description: Unauthorized
  /api/v1/harness/classify:
    post:
      tags: [Harness]
      summary: Classify a chat request without executing it
      description: >-
        Fork(harness B1). Stage-1 heuristics (free) classify the body into
        type (code, research, math, vision, search, chat), complexity (fast,
        deep) and detected modalities, plus the capability alias the harness
        would route to. With useModel=true a cheap classifier call may
        refine a low-confidence verdict; failures degrade to heuristics.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Task classification
        "400":
          description: Body lacks messages/prompt/input
        "401":
          description: Unauthorized
  /api/v1/harness/task:
    post:
      tags: [Harness]
      summary: Classify, route and execute a chat request in one call
      description: >-
        Fork(harness B1). The Gemini-style allocator as a single endpoint:
        the body is classified, its model rewritten to the classification's
        capability alias (a reserved name resolving to an ephemeral priority
        combo over the tag index's best specialists, with native failover),
        and executed on this server's /v1/chat/completions with the caller's
        credentials. Streaming passes through verbatim. Query params:
        alias= forces a route; classify_only=true returns the decision only.
        Response carries X-Harness-Route and X-Harness-Tier headers.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Routed chat response (or decision with classify_only)
        "400":
          description: Invalid body
        "401":
          description: Unauthorized
...
  /api/providers:
    get:
      tags: [Providers]
      summary: List provider connections
      responses:
        "200":
          description: Provider connection list
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProviderConnection"
    post:
      tags: [Providers]
      summary: Create provider connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderConnectionCreate"
      responses:
        "201":
          description: Created provider connection
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Two bearer families are accepted. Inference API keys (typically `sk-…`)
        authorize `/v1/*`. Management routes also accept `oma_live_…` Access Tokens
        (Settings → Access Tokens / `omniroute connect`) and API keys whose metadata
        includes `manage` or `admin` scope. See docs/guides/MANAGEMENT-AUTH.md.
        Bearer credentials are accepted on management routes that use this scheme;
        they are not rejected solely for being Bearer.
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie (auth_token) for protected management routes. Distinct from Bearer Access Tokens and API keys. See docs/guides/MANAGEMENT-AUTH.md.
...
  schemas:
...
    ProviderConnection:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        name:
          type: string
        url:
          type: string
        isActive:
          type: boolean
        maxConcurrent:
          type: integer
          nullable: true
          minimum: 0
        priority:
          type: integer
        testStatus:
          type: string
          enum: [active, error, untested]
        createdAt:
          type: string
          format: date-time
    ProviderConnectionCreate:
      type: object
      required: [provider, url]
      properties:
        provider:
          type: string
          example: openai
        name:
          type: string
        url:
          type: string
        apiKey:
          type: string
        isActive:
          type: boolean
          default: true
        maxConcurrent:
          type: integer
          nullable: true
          minimum: 0
...